adding test suit

+++ src/test/testCon.ts
@@ -1,16 +1,24 @@
-import { createConnection } from 'typeorm';
+import { CONFIG_DATABASE_PROVIDER, CONFIG_TEST_DATABASE_URL } from '@config/variables';
+import { Connection, createConnection } from 'typeorm';
 
-export const testConn = (drop: boolean = false) => {
-  return createConnection({
-    name: 'default',
-    type: 'mysql',
-    host: '127.0.0.1',
-    port: 3306,
-    username: 'Admin',
-    password: 'Admin1234!',
-    database: 'moon_shot_test',
-    synchronize: drop,
-    dropSchema: drop,
-    entities: [__dirname + '/../entities/*.*'],
-  });
+let client: Connection | null;
+
+export const testConn = async (drop: boolean = false): Promise<Connection> => {
+  if (!client) {
+    client = await createConnection({
+      name: 'default',
+      type: CONFIG_DATABASE_PROVIDER === 'postgres' ? 'postgres' : 'mysql',
+      url: CONFIG_TEST_DATABASE_URL,
+      synchronize: drop,
+      dropSchema: drop,
+      entities: [
+        __dirname + '/../entities/Recipes.entity.ts',
+        __dirname + '/../entities/Reservation.entity.ts',
+        __dirname + '/../entities/Restaurant.entity.ts',
+        __dirname + '/../entities/User.entity.ts',
+        __dirname + '/../entities/UserRole.entity.ts',
+      ],
+    });
+  }
+  return client;
 };
